interpd.cu : bug fix: CUDA ambiguous function call.

+++ src/interpd.cu
@@ -334,7 +334,7 @@ __device__ void wsinterpd_temp(T2 * __restrict__ y,
                 v += js * dstride[3 + 4*s]; // add pitched index for this dim (samples)
             }
             
-            const T2 a = isinf(tau[u]) ? no_v : (T2){cosf(omega * tau[u]), sinf(omega * tau[u])}; // modulation phasor
+            const T2 a = isinf((float)tau[u]) ? no_v : (T2){cosf(omega * tau[u]), sinf(omega * tau[u])}; // modulation phasor
             const T2 val = a * w[k] * sample(&x[(v)*T], (V)tau[u], flag, no_v, T); // weighted sample
             atomicAddStore(&y[l], val); // store
         }
@@ -387,7 +387,7 @@ __device__ void wsinterpd2_temp(T2 * __restrict__ y,
                 v += js * dstride[4 + 5*s]; // add pitched index for this dim (samples)
             }
             const U  t = tau1[r] + tau2[u]; // time
-            const T2 a = isinf(t) ? no_v : (T2){cosf(omega * t), sinf(omega * t)}; // modulation phasor
+            const T2 a = isinf((float)t) ? no_v : (T2){cosf(omega * t), sinf(omega * t)}; // modulation phasor
             const T2 val = a * w[k] * sample(&x[(v)*T], (V)t, flag, no_v, T); // weighted sample
             atomicAddStore(&y[l], val); // store
         }
